third_party/blink/web_tests/http/tests/origin_trials/webexposed/window-placement-origin-trial-interfaces.html

<!DOCTYPE html>
<meta charset="utf-8">
<!-- Generate token with the command:
generate_token.py http://127.0.0.1:8000 WindowPlacementOT2 --expire-timestamp=2000000000 --version 3
-->
<meta http-equiv="origin-trial" content="A2olqhso0OsTDx8t8ph298kXO0YKw7i06MsaVzw736c/wRA+nYG9eWETvsnNJxVa/P+R60Nxh3f5vCx1prSJnwQAAABaeyJvcmlnaW4iOiAiaHR0cDovLzEyNy4wLjAuMTo4MDAwIiwgImZlYXR1cmUiOiAiV2luZG93UGxhY2VtZW50T1QyIiwgImV4cGlyeSI6IDIwMDAwMDAwMDB9" />
<title>Window Placement API - interfaces exposed by origin trial</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/origin-trials-helper.js"></script>
<script src="/serviceworker/resources/test-helpers.js"></script>
<script>
'use strict'

test(t => {
  OriginTrialsHelper.check_properties_exist(this, {
     'global': ['getScreens'],
  });
}, 'Window Placement API interfaces and properties in Origin-Trial enabled document.');

test(t => {
  assert_true('getScreens' in window, 'getScreens function exists on window scope');
  assert_not_equals(window.getScreens, undefined, 'getScreens function can be accessed on window');
}, 'Window Placement API entry point in Origin-Trial enabled document.');

</script>
